Added Server Target and tested it
Added Server Target and locally verified that it is going to build.
Hit a snag, had to remove "binaries" and "intermediate" folder to make the build process to work otherwise it would just tell me that the build finished without errors after only 3 seconds instead of the usual 7min. Dont fully understand why the removal fixes the problem.